Added a very simplified, extremely basic, probably incomplete configuration file for cmake build systems.  As my knowledge of cmake is pretty limited, I didn't use any of the "special find" functions: for example, qxt is assumed to live  in C:\libQxt\ on Windows platforms, and in either /usr/local/Qxt or /usr/Qxt for everything else- sorry mac users!
I have tested this on 4 or 5 unix/linux systems and two windows (xp and 7) systems and
it essentially does what is expected: finds the proper include directories, includes them,
finds the libraries, and makes sure that they are used in linking.

This commit only adds the file.  Developers that wish to make use of it will
either need to pass -DQxt_DIR=/path/to/Qxt/features/QxtConfig.cmake, or
copy it:
cp /path/to/Qxt/features/QxtConfig.cmake /path/to/cmake/Modules/FindQxt.cmake

At some point the file can be more automagical like FindQt4.cmake/Qt4Uses.cmake,
but for now it works and is a (basic) solution that myself and a number of other
people using both cmake and libQxt will appreciate.